✅ test: parser 부분 테스트 정리
